requirements.txt
fastapi[standard]
boto3
python-dotenv